Enable hash traversal order randomization
This patch causes Perl to use a random order per hash when traversing hashes
via keys(), values() and each().

The random order is determined by add a new property xhv_rand to struct
xpvhv. This property is calculated on per traverse of the hash by using a
pointer/integer hash function on the address of the bucket array, and then
XORing it with the bucket index counter during traverse. This causes each
hash to have a different order for traversing the buckets.

The purpose of this change is to make hash seed discovery attacks harder.
An attacker loses a lot of information when the keys are "randomized".

The ptr_hash() sub is borrowed, with changes, from autobox, which is released
under the same terms as Perl.