randomize bucket split insertion in hsplit()
We use the HvRAND data as a set of flags that we rotate
over as we move items into the high buckets. If the flag is
set we insert one after the top, if the flag is not set we
insert into the top like we used to. This means the order
will be perturbed somewhat.