Ensure that Module::Pluggable returns plugins in a predictable order
As of 5.18 hash traversal order will vary per-hash, and indeed more
frequently. The likelihood this results in one of the tests in M::P
failing becomes very high.

In general the test expectations were fragile, but probably almost
never failed. With hash traversal randomization they will fail
regularly.

Note the use a temporary array in part of this patch is because
sort in scalar context is undefined.